+++ Configuration/TCA/Overrides/sys_template.php
@@ -1,4 +1,8 @@
 <?php
-defined('TYPO3') || die();
 
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile('simplequiz', 'Configuration/TypoScript', 'Simplequiz');
+defined('TYPO3') or die();
+
+call_user_func(function () {
+    $obj = new \Wacon\Simplequiz\Bootstrap\TCA\SysTemplate();
+    $obj->invoke();
+});