+++ ext_tables.php
@@ -1,7 +1,0 @@
-<?php
-defined('TYPO3') || die();
-
-(static function() {
-    // please insert your own code her
-})();
-## EXTENSION BUILDER DEFAULTS END TOKEN - Everything BEFORE this line is overwritten with the defaults of the extension builder