+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [dev-2.0.0-0]
-- [IMPORTANT] Rework Quiz model
+- [IMPORTANT] Rework all model to comply to the relational database model
+- [IMPORTANT] Rework TCA inclusion
 - [FEATURE] Change icons to WACON Logo
+- [FEATURE] Add Flexform for Simplequiz Plugin
+- [FEATURE] Rework ext_localconf and ext_tables.php
 
 ## [1.0.0] - 2024-03-31